requirements.txt
bcrypt                 
beautifulsoup4         
blinker                
certifi                
charset-normalizer     
click                  
colorama               
distlib                
dnspython              
email-validator        
exceptiongroup         
Faker                  
filelock               
Flask                  
Flask-Bcrypt           
Flask-Cors             
Flask-Login            
flask-marshmallow      
Flask-Principal        
Flask-Security-Too     
Flask-SQLAlchemy       
Flask-WTF              
greenlet               
gunicorn               
idna                   
importlib-resources    
iniconfig              
itsdangerous           
Jinja2                 
MarkupSafe             
marshmallow            
marshmallow-sqlalchemy
packaging
passlib
pip
platformdirs
playwright
pluggy
psycopg-binary
psycopg2
pyee
pyjokes
pytest
pytest-base-url
pytest-playwright
pytest-timeout
python-dateutil
python-dotenv
python-slugify
requests
setuptools
six
soupsieve
SQLAlchemy
text-unidecode
tomli
typing_extensions
urllib3
virtualenv
Werkzeug
WTForms